[LEMS-4205/fix-safari-point-label-jitter] address review comments
