+++ packages/perseus-editor/src/widgets/__docs__/interactive-graph-editor.stories.tsx
@@ -24,7 +24,6 @@ import {
     linearSystemWithStartingCoordsQuestion,
     linearWithStartingCoordsQuestion,
     pointQuestionWithStartingCoords,
-    pointWithCustomLabelQuestion,
     polygonWithStartingCoordsQuestion,
     quadraticWithStartingCoordsQuestion,
     rayWithStartingCoordsQuestion,
